(refactor): Changing directories
