+++ src/Infrastructure/EntryPoints/Console/Dice/RollDiceCommand.php
@@ -7,8 +7,8 @@ namespace RPGPlayground\Infrastructure\EntryPoints\Console\Dice;
 use Monolog\Level;
 use RPGPlayground\Application\UseCase\Dice\RollDice\RollDiceUseCase;
 use RPGPlayground\Application\UseCase\Dice\RollDice\RollDiceUseCaseInput;
-use RPGPlayground\Domain\ValueObjects\App\Dice;
-use RPGPlayground\Infrastructure\Handler\LogHandler;
+use RPGPlayground\Core\Handler\LogHandler;
+use RPGPlayground\Domain\ValueObjects\Dice;
 use Symfony\Component\Console\Attribute\AsCommand;
 use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Input\InputArgument;
